requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
contourpy==1.1.1
crispy-bootstrap5==0.7
cycler==0.12.1
Django==4.2.6
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-pagination==1.0.7
fonttools==4.43.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
